% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_statevars.R
\name{assign_ptable}
\alias{assign_ptable}
\title{Assign p table}
\usage{
assign_ptable(sv)
}
\arguments{
\item{sv}{df of s and n vals}
}
\value{
sv + column for which p table to use
}
\description{
Assign p table
}
