% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_statevars.R
\name{define_statevars}
\alias{define_statevars}
\title{Define statevars combinations}
\usage{
define_statevars(s_range = c(0, 6.5), n_range = c(0, 10.5),
  logs = TRUE, by = 0.5)
}
\arguments{
\item{s_range}{Range of S. Must be all positive numbers.}

\item{n_range}{Range of N. Must be all positive numbers. Max of n_range must > min of s_range.}

\item{logs}{are s and n ranges logs?}

\item{by}{by}
}
\value{
df of all allowable s and n combinations
}
\description{
Define statevars combinations
}
