% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fs.R
\name{sample_fs}
\alias{sample_fs}
\title{Sample the feasible set}
\usage{
sample_fs(s, n, nsamples, p_table = NULL)
}
\arguments{
\item{s}{s}

\item{n}{n}

\item{nsamples}{nb samples}

\item{p_table}{p table}
}
\value{
long dataframe of sim, rank, and abundance
}
\description{
Sample the feasible set
}
